Comparison Section
I added an idea/sketch of what we could do for the tables on the comparison section of the By Metro tab. Unfortunatley, I cannot get the inputs to work when I run the app locally so I can't test it. But the sandbox script has a functioning sample of what I want to have happen