Attempt to address error
When the user changes a category tehre is a moment where a "subscript out of bounds" error is displayed. I think that this is due to the fact that the national category resolves before the national variable is able to change. This commit simplifies the national map logic so that the error isn't on screen for a long period of time, btu does not address the underlying issue.