Update comp plot titles
There was code that renamed plot_1 and plot_2, reversing if it was with or without disabilities. This renaming was inaccurate, according to Joy and my own exploration of data.

A big picture problem here is that the code implicitly assumes that the left plot will be "with disabilities" and the right side "without," although this is actually tied to the census variables. The way to make this more accurate would be to add a binary variable column in the data dictionary, specifying if it is a variable that represents people with disabilities (1). This would be a good strategy long-term if we aim to further build out this project or if we are redoing the website.